+++ src/m_config.h
@@ -97,6 +97,7 @@ extern float    joy_deadzone_right;
 extern bool     joy_invertyaxis;
 extern int      joy_rumble_barrels;
 extern int      joy_rumble_damage;
+extern bool     joy_rumble_pickup;
 extern int      joy_rumble_weapons;
 extern float    joy_sensitivity_horizontal;
 extern float    joy_sensitivity_vertical;
@@ -505,6 +506,8 @@ enum
 #define joy_deadzone_right_max             100.0f
 
 #define joy_invertyaxis_default            false
+
+#define joy_rumble_pickup_default          true
 
 #define joy_rumble_damage_min              0
 #define joy_rumble_damage_default          100